Synthesize in iterations, play when first done
FluidSynth can be pretty CPU-intense on older hardware and even newer
hardware and so when transitioning to the end-screen or a new level, the
game used to freeze to let the music be synthesized. Even though 73492d8
moved this action to a different thread, the music was only played once
it was synthesized in its entirety. Now, synthesizing is done in chunks
and once the first one is done, it is played immediately and new chunks
are queued to OpenAL.